+++ jkorm/src/main/kotlin/com/tang/jkorm/wrapper/statement/ComparisonStatement.kt
@@ -1,0 +1,17 @@
+package com.tang.jkorm.wrapper.statement
+
+import com.tang.jkorm.wrapper.enumeration.ComparisonOperator
+
+/**
+ * Comparison statement
+ *
+ * @author Tang
+ */
+class ComparisonStatement(val column: String, val value: Any, val comparisonOperator: ComparisonOperator) {
+
+    fun appendSql(sql: StringBuilder, parameters: MutableList<Any?>) {
+        parameters.add(value)
+        sql.append("$column${comparisonOperator.value}?")
+    }
+
+}