+++ CHANGELOG.md
@@ -42,3 +42,4 @@
  * Add `OrderItem` for paginate order by
  * Add `OrderItem` for paginate order by with generics
  * Add property reference in `OrderItem`
+ * Add `UpdateWrapper` for update statement